# Conflicts:
#	src/main/java/cn/ruanyun/backInterface/modules/business/itemAttrKey/serviceimpl/IItemAttrKeyServiceImpl.java

+++ src/main/java/cn/ruanyun/backInterface/modules/business/grade/pojo/Grade.java
@@ -27,6 +27,6 @@ public class Grade extends RuanyunBaseEntity {
     /**
      * 星星数
      */
-    private Integer startLevel;
+    private Double startLevel;
 
 }
